+++ lib/transforms/map.ts
@@ -1,7 +1,9 @@
 import { sanitizeOptions } from "../utils/sanitizeOptions.ts";
 import { ChunkTransformFunc } from "../types.ts";
-export type MapOptions = {};
-const defaults: MapOptions = {};
+import { z, ZodSchema } from "../../deps/zod.ts";
+export type MapOptions<T> = {
+  validator?: ZodSchema<T>;
+};
 
 /**
  * Creates a TransformStream (Webstream) which converts from one chunk format to another.
@@ -10,13 +12,24 @@ const defaults: MapOptions = {};
  */
 export function map<T, R>(
   mapFunc: ChunkTransformFunc<T, R>,
-  options?: MapOptions,
+  options?: MapOptions<T>,
 ) {
-  const {} = sanitizeOptions(options, defaults);
+  const defaults: MapOptions<T> = {};
+  const { validator } = sanitizeOptions(options, defaults);
 
   const transform = new TransformStream<T, R>({
     start: () => {},
     transform(chunk, controller) {
+      if (validator) {
+        const result = validator.safeParse(chunk);
+        if (!result.success) {
+          controller.error(result.error);
+          return;
+        }
+
+        chunk = result.data;
+      }
+
       controller.enqueue(mapFunc(chunk));
     },
   });